Add captions styling menu, clean up architecture
