+++ src/css/controls/imports/settings-menu.less
@@ -21,7 +21,7 @@
         height: 35%;
         width: 25%;
 
-        .jw-settings-topbar {
+        .jw-settings-topbar:not(.jw-nested-menu-open) {
             .jw-icon-inline {
                 .square(60px);
 
@@ -35,6 +35,10 @@
                     }
                 }
             }
+        }
+
+        .jw-settings-back {
+            min-width: 60px;
         }
     }
 
@@ -76,6 +80,10 @@
             }
         }
     }
+
+    .jw-settings-reset {
+        text-decoration: underline;
+    }
 }
 
 .jw-settings-topbar {
@@ -88,30 +96,44 @@
 
     &.jw-nested-menu-open {
         padding: 0;
+
         .jw-icon:not(.jw-settings-close):not(.jw-settings-back) {
             display: none;
         }
+
         .jw-svg-icon-close {
             width: 20px;
         }
+
         .jw-svg-icon-arrow-left {
             height: 12px;
         }
+
         .jw-settings-topbar-text {
             display: block;
         }
+    }
+
+    .jw-settings-back {
+        min-width: 44px;
+    }
+
+    .jw-settings-topbar-buttons {
+        display: inherit;
+        width: 100%;
+        height: 100%;
     }
 
     .jw-settings-topbar-text {
         display: none;
         color: #fff;
         font-size: 13px;
+        width: 100%;
     }
 
     .jw-settings-close {
         margin-left: auto;
     }
-
 }
 
 .jw-settings-submenu {
@@ -121,7 +143,7 @@
     overflow-y: auto;
     padding: 8px 20px 0 5px;
     .scrollbar-style();
-    
+
     &.jw-settings-submenu-active {
         display: block;
     }
@@ -130,6 +152,7 @@
         box-shadow: 0 2px 9px 0 #1d1d1d;
         background-color: #2f2d2d;
         margin: -8px -20px 0 -5px;
+
         .jw-settings-content-item {
             cursor: pointer;
             text-align: right;
@@ -138,10 +161,15 @@
         }
     }
 
-    .jw-settings-content-value-wrapper {
+    .jw-settings-value-wrapper {
         float: right;
         display: flex;
         align-items: center;
+
+        .jw-settings-content-item-arrow {
+            display: flex;
+        }
+
         .jw-svg-icon-arrow-right {
             width: 8px;
             margin-left: 5px;
